+++ projects/angular-material-extensions/select-country/src/lib/tokens/index.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from '@angular/core';
+
+export const MatSelectCountryLangToken = new InjectionToken<string>('MatSelectCountryLangToken');