+++ src/app/app.component.ts
@@ -1,7 +1,8 @@
-import {Component, OnInit} from '@angular/core';
-import {Country} from '@angular-material-extensions/select-country';
-import {Angulartics2GoogleAnalytics} from 'angulartics2/ga';
-import {FormBuilder, FormControl, FormGroup} from '@angular/forms';
+import { Component, OnInit } from '@angular/core';
+import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+
+import { Country } from '@angular-material-extensions/select-country';
+import { Angulartics2GoogleAnalytics } from 'angulartics2/ga';
 
 @Component({
   selector: 'app-root',
@@ -9,29 +10,89 @@ import {FormBuilder, FormControl, FormGroup} from '@angular/forms';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements OnInit {
-
-  title = 'select-county';
+  title = 'select-country';
 
   countryFormControl = new FormControl();
+  countryRequiredFormControl = new FormControl(null, [Validators.required]);
   countryFormGroup: FormGroup;
 
-  constructor(angulartics2GoogleAnalytics: Angulartics2GoogleAnalytics,
-              private formBuilder: FormBuilder) {
+  defaultValue: Country = {
+    name: 'Deutschland',
+    alpha2Code: 'DE',
+    alpha3Code: 'DEU',
+    numericCode: '276',
+    callingCode: '+49'
+  };
+
+  predefinedCountries: Country[] = [
+    {
+      name: 'Germany',
+      alpha2Code: 'DE',
+      alpha3Code: 'DEU',
+      numericCode: '276',
+      callingCode: '+49'
+    },
+    {
+      name: 'Greece',
+      alpha2Code: 'GR',
+      alpha3Code: 'GRC',
+      numericCode: '300',
+      callingCode: '+30'
+    },
+    {
+      name: 'France',
+      alpha2Code: 'FR',
+      alpha3Code: 'FRA',
+      numericCode: '250',
+      callingCode: '+33'
+    },
+    {
+      name: 'Belgium',
+      alpha2Code: 'BE',
+      alpha3Code: 'BEL',
+      numericCode: '056',
+      callingCode: '+32'
+    }
+  ];
+
+  constructor(
+    angulartics2GoogleAnalytics: Angulartics2GoogleAnalytics,
+    private formBuilder: FormBuilder
+  ) {
     angulartics2GoogleAnalytics.startTracking();
   }
 
-  onCountrySelected($event: Country) {
-    console.log($event);
-  }
-
   ngOnInit(): void {
-
     this.countryFormGroup = this.formBuilder.group({
-      country: []
+      country: [
+        {
+          value: {
+            name: 'Deutschland',
+            alpha2Code: 'DE',
+            alpha3Code: 'DEU',
+            numericCode: '276',
+            callingCode: '+49'
+          },
+          disabled: false
+        }
+      ]
     });
 
-    this.countryFormGroup.get('country').valueChanges.subscribe(country => console.log('this.countryFormGroup.get(\'country\').valueChanges', country));
+    this.countryFormGroup
+      .get('country')
+      .valueChanges.subscribe((country) =>
+      console.log(
+        'this.countryFormGroup.get(\'country\').valueChanges',
+        country
+      )
+    );
 
-    this.countryFormControl.valueChanges.subscribe(country => console.log('this.countryFormControl.valueChanges', country));
+    this.countryFormControl.valueChanges.subscribe((country) =>
+      console.log('this.countryFormControl.valueChanges', country)
+    );
+  }
+
+  onCountrySelected($event: Country): void {
+    console.log($event);
   }
 }