+++ src/app/app.server.module.ts
@@ -1,16 +1,16 @@
 import { NgModule } from '@angular/core';
 import { ServerModule } from '@angular/platform-server';
+import { FlexLayoutServerModule } from '@angular/flex-layout/server';
 
 import { AppModule } from './app.module';
 import { AppComponent } from './app.component';
-import {FlexLayoutServerModule} from '@angular/flex-layout/server';
 
 @NgModule({
   imports: [
     AppModule,
     FlexLayoutServerModule,
     ServerModule
-],
+  ],
   bootstrap: [AppComponent],
 })
 export class AppServerModule {}